LayoutTests/css3/touch-action/touch-action-parsing-expected.txt

PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('touch-action') is 'manipulation'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('touch-action') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
Test the parsing of touch-action.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
